add page components
