+++ src/app/layout.tsx
@@ -1,11 +1,6 @@
 "use client";
 
 import { ConfigWrappers } from "@/components/ConfigWrappers";
-import { siteTitle } from "@/constants";
-
-const title = siteTitle;
-const description = "An optimistic oracle built for web3";
-const images = "/assets/twitter-card.png";
 
 export default function RootLayout({
   children,
