+++ src/app/page.tsx
@@ -1,10 +1,5 @@
-"use client";
+import Verify from "@/components/pages/verify";
 
-import { OracleQueries } from "@/components";
-import { usePage } from "@/hooks";
-
-export default function Verify() {
-  const page = "verify";
-  usePage(page);
-  return <OracleQueries page={page} />;
+export default function VerifyPage() {
+  return <Verify />;
 }
