+++ src/app/settled/page.tsx
@@ -1,10 +1,5 @@
-"use client";
+import Settled from "@/components/pages/settled";
 
-import { OracleQueries } from "@/components";
-import { usePage } from "@/hooks";
-
-export default function Settled() {
-  const page = "settled";
-  usePage(page);
-  return <OracleQueries page={page} />;
+export default function SettledPage() {
+  return <Settled />;
 }
