+++ src/components/pages/propose.tsx
@@ -1,0 +1,10 @@
+"use client";
+
+import { OracleQueries } from "@/components";
+import { usePage } from "@/hooks";
+
+export default function Propose() {
+  const page = "propose";
+  usePage(page);
+  return <OracleQueries page={page} />;
+}
