+++ src/components/pages/settled.tsx
@@ -1,0 +1,10 @@
+"use client";
+
+import { OracleQueries } from "@/components";
+import { usePage } from "@/hooks";
+
+export default function Settled() {
+  const page = "settled";
+  usePage(page);
+  return <OracleQueries page={page} />;
+}
