+++ src/components/pages/verify.tsx
@@ -1,0 +1,10 @@
+"use client";
+
+import { OracleQueries } from "@/components";
+import { usePage } from "@/hooks";
+
+export default function Verify() {
+  const page = "verify";
+  usePage(page);
+  return <OracleQueries page={page} />;
+}
